refactor: extract settings-related state into new settingsStore
